+++ backend/storj/object.go
@@ -1,7 +1,7 @@
 //go:build !plan9
 // +build !plan9
 
-package tardigrade
+package storj
 
 import (
 	"context"
