+++ backend/storj/storj_test.go
@@ -1,8 +1,8 @@
 //go:build !plan9
 // +build !plan9
 
-// Test Tardigrade filesystem interface
-package tardigrade_test
+// Test Storj filesystem interface
+package storj_test
 
 import (
 	"github.com/rclone/rclone/backend/storj"
@@ -14,7 +14,7 @@ import (
 // TestIntegration runs integration tests against the remote
 func TestIntegration(t *testing.T) {
 	fstests.Run(t, &fstests.Opt{
-		RemoteName: "TestTardigrade:",
+		RemoteName: "TestStorj:",
 		NilObject:  (*storj.Object)(nil),
 	})
 }
