+++ backend/storj/storj_unsupported.go
@@ -1,0 +1,4 @@
+//go:build plan9
+// +build plan9
+
+package storj
