+++ cmd/out.csv
@@ -1,0 +1,3 @@
+,
+,
+Bytes,Char
