+++ docs/layouts/chrome/navbar.html
@@ -95,6 +95,7 @@
           <a class="dropdown-item" href="/putio/"><i class="fas fa-parking"></i> put.io</a>
           <a class="dropdown-item" href="/seafile/"><i class="fa fa-server"></i> Seafile</a>
           <a class="dropdown-item" href="/sftp/"><i class="fa fa-server"></i> SFTP</a>
+          <a class="dropdown-item" href="/storj/"><i class="fa fa-server"></i> Storj DCS</a>
           <a class="dropdown-item" href="/sugarsync/"><i class="fas fa-dove"></i> SugarSync</a>
           <a class="dropdown-item" href="/tardigrade/"><i class="fas fa-dove"></i> Tardigrade</a>
           <a class="dropdown-item" href="/uptobox/"><i class="fa fa-archive"></i> Uptobox</a>
